Combine DF's across pacing directions within a timepoint

Combine DF's across pacing directions within a timepoint

combine_rds_pacing_dir_v2(
  id,
  df,
  grouping,
  base_dir = "data/",
  suffix = "_full",
  step,
  timepoint,
  wavefront
)

Arguments

id

Subject id

df

Data frame containing data from the pacing direction that you want to merge

grouping

"carto_mesh" or "lv_volt" or other group that you've filtering within

base_dir

Directory where you're saving this data to

suffix

Any suffix you want to add to the data file name, name looks like step2_t2_rvpaced_full.rds

step

What step in your process are you in?

timepoint

The timepoint that you're currently working with

wavefront

The pacing wavefront that you're currently merging in

Value

Creates a file in wherever we declared our base_dir